modified default read of mangoid
it's commented out but it addresses an issue
where a tester had it complain about no primary_preferred
server. This _should_ be attempting to connect to
a primary server (no preferred).

if folks need this they can uncomment it.
it's been documented in the troubleshooting guide on the site

I'm not uncommenting it because I can't replicate
the problem myself, so I'm not sure if it
actually helps or hurts anything